+++ packages/react-multi-crop/src/ReactMultiCrop/ReactMultiCrop.stories.js
@@ -1,22 +1,34 @@
-import React from "react";
-import ReactMultiCrop from "./ReactMultiCrop";
+import React from 'react';
+import ReactMultiCrop from './ReactMultiCrop';
 
 export default {
-  title: "ReactMultiCrop",
+  title: 'ReactMultiCrop',
   component: ReactMultiCrop,
 };
 
 const Template = (args) => <ReactMultiCrop hideLabel {...args} />;
 export const Default = Template.bind({});
 Default.args = {
-  image: "https://picsum.photos/200",
-  hideLabel: true,
+  image: 'https://picsum.photos/200',
+  showLabel: false,
+  showButton: false,
+  includeDataUrl: false,
+  includeHtmlCanvas: false,
+  cropBackgroundColor: 'yellow',
+  cropBackgroundOpacity: 0.3,
+  cropOutlineColor: 'black',
+  cropOutlineWidth: 3,
+  input: {
+    onChange(value) {
+      console.log(value);
+    },
+  },
 };
 
 Default.parameters = {
   docs: {
     description: {
-      component: "Example load an image to cropper",
+      component: 'Example load an image to cropper',
     },
     source: {
       code: `<ReactMultiCrop image="https://picsum.photos/200" hideLabel />`,
@@ -26,9 +38,9 @@ Default.parameters = {
 
 Default.argTypes = {
   image: {
-    description: "image url",
+    description: 'image url',
   },
   hideLabel: {
-    description: "hide the default label",
+    description: 'hide the default label',
   },
 };