+++ packages/react-multi-crop/src/ReactMultiCrop/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./ReactMultiCrop";
+export { default } from './ReactMultiCrop';